+++ sentry/src/main/java/io/sentry/ITransactionProfiler.java
@@ -8,12 +8,17 @@ import org.jetbrains.annotations.Nullable;
 /** Used for performing operations when a transaction is started or ended. */
 @ApiStatus.Internal
 public interface ITransactionProfiler {
-  void onTransactionStart(@NotNull ITransaction transaction);
+  boolean isRunning();
+
+  void start();
+
+  void bindTransaction(@NotNull ITransaction transaction);
 
   @Nullable
   ProfilingTraceData onTransactionFinish(
       @NotNull ITransaction transaction,
-      @Nullable List<PerformanceCollectionData> performanceCollectionData);
+      @Nullable List<PerformanceCollectionData> performanceCollectionData,
+      @NotNull SentryOptions options);
 
   /** Cancel the profiler and stops it. Used on SDK close. */
   void close();