+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -454,6 +454,16 @@ public class SentryOptions {
 
   @ApiStatus.Experimental private boolean enableBackpressureHandling = false;
 
+  /** Whether to enable startup profiling, depending on profilesSampler or profilesSampleRate. */
+  private boolean enableStartupProfiling = false;
+
+  /**
+   * Profiling traces rate. 101 hz means 101 traces in 1 second. Defaults to 101 to avoid possible
+   * lockstep sampling. More on
+   * https://stackoverflow.com/questions/45470758/what-is-lockstep-sampling
+   */
+  private int profilingTracesHz = 101;
+
   /**
    * Adds an event processor
    *
@@ -735,6 +745,20 @@ public class SentryOptions {
     }
 
     return dsnHash != null ? new File(cacheDirPath, dsnHash).getAbsolutePath() : cacheDirPath;
+  }
+
+  /**
+   * Returns the cache dir path if set, without the appended dsn hash.
+   *
+   * @return the cache dir path, without the appended dsn hash, or null if not set.
+   */
+  @Nullable
+  String getCacheDirPathWithoutDsn() {
+    if (cacheDirPath == null || cacheDirPath.isEmpty()) {
+      return null;
+    }
+
+    return cacheDirPath;
   }
 
   /**
@@ -1627,13 +1651,17 @@ public class SentryOptions {
   }
 
   /**
-   * Sets the listener interface to perform operations when a transaction is started or ended.
+   * Sets the listener interface to perform operations when a transaction is started or ended. It
+   * only has effect if no profiler was already set.
    *
    * @param transactionProfiler - the listener for operations when a transaction is started or ended
    */
   public void setTransactionProfiler(final @Nullable ITransactionProfiler transactionProfiler) {
-    this.transactionProfiler =
-        transactionProfiler != null ? transactionProfiler : NoOpTransactionProfiler.getInstance();
+    // We allow to set the profiler only if it was not set before, and we don't allow to unset it.
+    if (this.transactionProfiler == NoOpTransactionProfiler.getInstance()
+        && transactionProfiler != null) {
+      this.transactionProfiler = transactionProfiler;
+    }
   }
 
   /**
@@ -2126,6 +2154,25 @@ public class SentryOptions {
   }
 
   /**
+   * Whether to enable startup profiling, depending on profilesSampler or profilesSampleRate.
+   * Depends on {@link SentryOptions#isProfilingEnabled()}
+   *
+   * @return true if startup profiling should be started.
+   */
+  public boolean isEnableStartupProfiling() {
+    return isProfilingEnabled() && enableStartupProfiling;
+  }
+
+  /**
+   * Whether to enable startup profiling, depending on profilesSampler or profilesSampleRate.
+   *
+   * @param enableStartupProfiling true if startup profiling should be started.
+   */
+  public void setEnableStartupProfiling(boolean enableStartupProfiling) {
+    this.enableStartupProfiling = enableStartupProfiling;
+  }
+
+  /**
    * Whether to send modules containing information about versions.
    *
    * @param sendModules true if modules should be sent.
@@ -2216,6 +2263,22 @@ public class SentryOptions {
   @ApiStatus.Experimental
   public void setEnableBackpressureHandling(final boolean enableBackpressureHandling) {
     this.enableBackpressureHandling = enableBackpressureHandling;
+  }
+
+  /**
+   * Returns the rate the profiler will sample rates at. 100 hz means 100 traces in 1 second.
+   *
+   * @return Rate the profiler will sample rates at.
+   */
+  @ApiStatus.Internal
+  public int getProfilingTracesHz() {
+    return profilingTracesHz;
+  }
+
+  /** Sets the rate the profiler will sample rates at. 100 hz means 100 traces in 1 second. */
+  @ApiStatus.Internal
+  public void setProfilingTracesHz(final int profilingTracesHz) {
+    this.profilingTracesHz = profilingTracesHz;
   }
 
   @ApiStatus.Experimental
@@ -2311,7 +2374,8 @@ public class SentryOptions {
    *
    * @return SentryOptions
    */
-  static @NotNull SentryOptions empty() {
+  @ApiStatus.Internal
+  public static @NotNull SentryOptions empty() {
     return new SentryOptions(true);
   }